rename api users to authentication
