+++ src/pages/register/types.ts
@@ -1,4 +1,4 @@
-import { RegisterFormData } from "../../api/users/types";
+import { RegisterFormData } from "../../api/authentication/types";
 
 export type RegisterFormErrors = Partial<
     Record<keyof RegisterFormData, string>
